+++ liqwid-script-export/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
+## 2.4.0 -- 2024-06-04
+
+### Added
+
+- `mkTwoArgumentScriptInfo` will replace `mkPolicyInfo` and `mkStakeValidatorInfo`
+- `mkThreeArgumentScriptInfo` will replace `mkValidatorInfo`
+
 ## 2.3.0 -- 2023-01-11
 
 ### Added